requirements/local.txt
-r base.txt
# A configurable set of panels that display various debug information about the current
# request/response.
# http://django-debug-toolbar.readthedocs.org/
django-debug-toolbar==1.4

# The Python WSGI Utility Library
# http://werkzeug.pocoo.org/
Werkzeug==0.11.2

# Code coverage measurement for Python
# https://coverage.readthedocs.org/en/coverage-4.0.2/
coverage==4.0.3

# A plugin for py.test that provides a set of useful tools for testing Django
# applications and projects.
# https://pytest-django.readthedocs.org/
pytest-django==2.9.1

# A mature full-featured Python testing tool
# http://pytest.org/latest/
pytest==2.8.3

# Pytest plugin for measuring coverage.
# https://pypi.python.org/pypi/pytest-cov
pytest-cov==2.2.0

# the modular source code checker: pep8, pyflakes and co
# https://pypi.python.org/pypi/flake8/2.5.4
flake8==2.5.4

# Scrapy extension to write scraped items using Django models
# https://github.com/scrapy-plugins/scrapy-djangoitem
scrapy-djangoitem==1.1.1

# Scrapy is an application framework for crawling web sites
# and extracting structured data which can be used for a wide
# range of useful applications.
# http://doc.scrapy.org/en/1.1/
Scrapy==1.1.0

# Turn HTML into equivalent Markdown-structured text.
# https://pypi.python.org/pypi/html2text
html2text==2016.5.29

# Lightweight SOAP client
# https://pypi.python.org/pypi/suds
suds==0.4

